Move the ID structs into the metadata module
