+++ embassy-executor/src/metadata.rs
@@ -207,3 +207,23 @@ impl Metadata {
         self.deadline.set(instant_ticks);
     }
 }
+
+/// A unique ID for a task
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct TaskId(pub(crate) usize);
+impl TaskId {
+    /// Get the id as a raw number
+    pub fn get(&self) -> usize {
+        self.0
+    }
+}
+
+/// A unique ID for an executor
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct ExecutorId(pub(crate) usize);
+impl ExecutorId {
+    /// Get the id as a raw number
+    pub fn get(&self) -> usize {
+        self.0
+    }
+}
